@@ -3,20 +3,30 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "react": "^16.8.6",
-    "react-dom": "^16.4.2",
+    "react": "^17.0.2",
+    "react-dom": "^17.0.2",
     "react-helmet": "^6.1.0",
-    "react-router-dom": "^5.0.1"
+    "react-router-dom": "^6.2.2"
   },
   "devDependencies": {
     "react-scripts": "^5.0.0",
-    "react-snapshot": "^1.3.0"
+    "react-snap": "^1.23.0",
+    "serve": "^13.0.2"
   },
   "scripts": {
     "start": "react-scripts start",
-    "build": "react-scripts build && react-snapshot",
+    "build": "react-scripts build",
+    "postbuild": "react-snap",
+    "serve": "serve -s build",
     "test": "react-scripts test --env=jsdom",
     "eject": "react-scripts eject"
+  },
+  "reactSnap": {
+    "source": "build",
+    "minifyHtml": {
+      "collapseWhitespace": false,
+      "removeComments": false
+    }
   },
   "browserslist": {
     "production": [
