@@ -11,10 +11,10 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
-    "lodash": "^4.17.11"
+    "lodash": "^4.17.21"
   },
   "devDependencies": {
-    "eslint": "^6.0.1",
-    "jest": "^24.8.0"
+    "eslint": "^8.11.0",
+    "jest": "^27.5.1"
   }
 }
