@@ -10,6 +10,6 @@
   "author": "",
   "license": "ISC",
   "devDependencies": {
-    "electron": "^4.0.1"
+    "electron": "^17.1.2"
   }
 }
