@@ -4,7 +4,7 @@
   "description": "",
   "private": true,
   "scripts": {
-    "start": "webpack serve --mode=development",
+    "start": "parcel index.html -p 3000",
     "test": "jest",
     "lint": "eslint --fix ."
   },
@@ -12,16 +12,18 @@
   "license": "ISC",
   "devDependencies": {
     "dotenv-webpack": "^7.1.0",
-    "eslint": "^8.9.0",
+    "eslint": "^8.11.0",
     "eslint-config-airbnb-base": "^15.0.0",
     "eslint-plugin-import": "^2.25.4",
     "jest": "^27.5.1",
-    "webpack": "^5.69.1",
+    "parcel": "^2.4.0",
+    "process": "^0.11.10",
+    "webpack": "^5.70.0",
     "webpack-cli": "^4.9.2",
     "webpack-dev-server": "^4.7.4"
   },
   "dependencies": {
-    "axios": "^0.26.0",
+    "axios": "^0.26.1",
     "buffer": "^6.0.3",
     "sumi-hash": "^1.0.1",
     "tweetnacl": "^1.0.3"
