@@ -1,5 +1,4 @@
-import React from 'react';
-import { render } from 'react-snapshot';
+import { render } from "react-dom";
 
 import './index.css';
 
