@@ -1,9 +1,18 @@
+const path = require('path')
 const { app, BrowserWindow, ipcMain } = require('electron');
 
+console.log('main!');
+
 app.on('ready', () => {
-  const win = new BrowserWindow({ width: 800, height: 400 });
+  const win = new BrowserWindow({
+    width: 800,
+    height: 400,
+    webPreferences: {
+      preload: path.join(__dirname, 'preload.js'),
+    },
+  });
+
   win.loadFile('index.html');
-  console.log('main!');
 });
 
 ipcMain.on('handle-click', (event, arg) => {
